fix: shell control-flow block delimiters and bracket spacing
- Extract shared shell_block_open_for/shell_block_close_for into shell_syntax.rs
- Zsh: add close_on_transition: false, delegate to shared helpers
- Bash: delegate block_open_for/block_close_for to shared helpers
- Macro: add inner spaces for shell bracket groups ([ ... ] and [[ ... ]])
- Update golden files with corrected output
