+++ test-goldens/bash/macro_control_flow.bash
@@ -1,4 +1,4 @@
-if [$x -gt 0]; then
+if [ $x -gt 0 ]; then
     echo "positive"
 else
     echo "negative"
